refactor(#208): introduce utils for custom translators
